+++ 語料庫/widgets/目標音檔欄.py
@@ -1,0 +1,9 @@
+
+class 目標音檔欄():
+    def 目標音檔(self, obj):
+        return ('''<audio controls>
+        <source src='/音檔/{}/{}/{}/audio.wav'>
+        Your browser does not support the audio element.</audio>'''.format(
+            obj.音檔.id, obj.聲音開始時間, obj.聲音結束時間
+        ))
+    目標音檔.allow_tags = True
